requirements.txt
python-dotenv==1.0.1
random2==1.0.1 